% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyURL.R
\name{shinyURL.Server}
\alias{shinyURL.Server}
\alias{shinyURL.UI}
\title{Deprecated functions}
\usage{
shinyURL.Server(...)

shinyURL.UI(...)
}
\arguments{
\item{...}{Arguments passed to the new methods \code{\link{shinyURL.server}}
and \code{\link{shinyURL.ui}}.}
}
\description{
Deprecated functions
}

