% Generated by roxygen2 (4.0.2): do not edit by hand
\name{encodeShinyURL}
\alias{encodeShinyURL}
\title{Encode current application state into URL}
\usage{
encodeShinyURL(session, input)
}
\arguments{
\item{session}{parameter passed from the shiny "server function"}

\item{input}{parameter passed from the shiny "server function"}
}
\description{
Encode current application state into URL
}

