% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{shinyURL.Server}
\alias{shinyURL.Server}
\alias{shinyURL.UI}
\title{Deprecated functions in package \sQuote{shinyURL}}
\usage{
shinyURL.Server(...)

shinyURL.UI(...)
}
\arguments{
\item{...}{Arguments passed to the new methods}
}
\description{
These functions are provided for compatibility with older versions of
\sQuote{shinyURL} only, and will be defunct at the next release.
}
\details{
The following functions are deprecated and will be made defunct; use
  the replacement indicated below. \itemize{ \item{shinyURL.Server:
  \code{\link{shinyURL.server}}} \item{shinyURL.UI:
  \code{\link{shinyURL.ui}}} }
}

