% Generated by roxygen2 (4.0.2): do not edit by hand
\name{shinyURL}
\alias{shinyURL}
\title{Save and restore the state of a Shiny App}
\usage{
shinyURL(session, inputId = ".url")
}
\arguments{
\item{session}{parameter passed from the \code{shinyServer} function}

\item{inputId}{id of the text field containing the encoded URL}
}
\description{
Encode the state of a Shiny application into an URL and use URL parameters to initialize the app.
}

