% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyURL.R
\name{encodeShinyURL}
\alias{encodeShinyURL}
\title{Encode current application state into URL}
\usage{
encodeShinyURL(session)
}
\arguments{
\item{session}{parameter passed from the \code{shinyServer} function}
}
\description{
Encode current application state into URL
}
\note{
This function has been deprecated. Please use \code{\link{shinyURL}} instead.
}

