% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initFromURL}
\alias{initFromURL}
\title{Initialize application state from URL}
\usage{
initFromURL(session, nestedDependency = FALSE, self, encode, resume = NULL)
}
\arguments{
\item{session}{parameter passed from the \code{shinyServer} function}

\item{nestedDependency}{set to \code{TRUE} if you use reactive UI}

\item{self}{self-reference}

\item{encode}{reference to the URL encoder object}

\item{resume}{list of observers which start in a suspended state and need to be resumed after initialization finishes}
}
\description{
Initialize application state from URL
}
\note{
This function has been deprecated. Please use \code{\link{shinyURL}} instead.
}

