% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R, R/shinyURL.R, R/ui.R
\name{shinyURL.server}
\alias{shinyURL}
\alias{shinyURL.server}
\alias{shinyURL.ui}
\title{Save and restore the view state of a Shiny app}
\usage{
shinyURL.server(session, options)

shinyURL.ui(display = TRUE, label = "Share URL", width = "100\%",
  copyURL = TRUE, tinyURL = TRUE, ZeroClipboard.swf)
}
\arguments{
\item{session}{Typically the same as the optional parameter passed into the 
Shiny server function as an argument; if missing defaults to 
\code{getDefaultReactiveDomain()}}

\item{options}{Named list of options}

\item{display}{logical, should the shinyURL widget be displayed}

\item{label}{Label for the URL field}

\item{width}{The width of the URL text field, e.g. \code{'100\%'}, or 
\code{'400px'}; see \code{\link[shiny]{validateCssUnit}}.}

\item{copyURL}{Include a \sQuote{Copy} button for convenient copying to 
clipboard}

\item{tinyURL}{Use the TinyURL web service for shortening the URL}

\item{ZeroClipboard.swf}{URL of the file ZeroClipboard.swf, as 
passed to the \sQuote{swfPath} parameter of \sQuote{ZeroClipboard.config}; 
if missing defaults to the local copy distributed with shinyURL}
}
\value{
\code{shinyURL.server} returns a reactive expression evaluating to 
  the app's URL.
}
\description{
Encode the state of Shiny app's widgets into an URL query string, and use 
parameters from the URL query string to initialize the app.
}
\details{
The \code{shinyURL.server} method contains server logic for encoding
  and restoring the widgets' values. It is called from inside the app's 
  server script, and can take the \code{session} objects as argument.
  
  The argument \code{options} can contain a named list of options. These are
  set by a call to \code{\link[base]{options}} as \sQuote{shinyURL.name}. See below for a list of available options.

The \code{shinyURL.ui} widget consists of a text field containing an
  URL to the app's current view state.  By default it also features the 
  convenience \sQuote{Copy} button for copying the URL to clipboard, and a 
  \sQuote{TinyURL} button for querying the URL shortening web service.  The 
  inclusion of these buttons is optional and can be controlled by the 
  \code{copyURL} and \code{tinyURL} arguments, respectively.
  
  The \sQuote{Copy} feature uses the ZeroClipboard library, which provides an
  easy way to copy text to the clipboard using an invisible Adobe Flash movie
  and JavaScript. shinyURL includes the JavaScript code to your app 
  automatically, but you also need to have the \dQuote{ZeroClipboard.swf} 
  available to the browser. By default shinyURL uses a local copy distributed
  with the package; you can override this by setting the
  \code{ZeroClipboard.swf} argument to \code{shinyURL.ui}, for example, use 
  "//cdnjs.cloudflare.com/ajax/libs/zeroclipboard/2.2.0/ZeroClipboard.swf"
  for a file hosted on jsDelivr CDN.
}
\section{ShinyURL options}{
  
\describe{
 \item{\code{debug = TRUE}}{Print debug messages to the console}
}
}

\section{Quick setup}{
 To start using shinyURL in your Shiny app, follow these
  three steps: \enumerate{ \item Load the package in both 'server.R' an
  ui.R': \code{library("shinyURL")} \item Add a call to \code{ 
  shinyURL.server()} inside the server function \item Add the 
  \code{shinyURL.ui()} widget to the user interface}
}
\examples{
if (interactive()) {
  library("shiny")
  
  ## A Simple Shiny App
   
  shinyApp(
    ui = fluidPage(
      titlePanel("Hello Shiny!"),
      sidebarLayout(
        sidebarPanel(
          sliderInput("bins", "Number of bins:", min = 1, max = 50, value = 30),
          shinyURL.ui()
        ),
        mainPanel(
          plotOutput("plot")
        )
      )
    ),
    server = function(input, output, session) {
      shinyURL.server(session)
      output$plot <- renderPlot({
        x <- faithful[, 2]
        bins <- seq(min(x), max(x), length.out = input$bins + 1)
        hist(x, breaks = bins, col = 'darkgray', border = 'white')
      })
    }
  )

  ## Shiny Widgets Demo
  shinyAppDir( system.file('examples', 'widgets', package='shinyURL') )

  ## Tabsets Demo
  shinyAppDir( system.file('examples', 'tabsets', package='shinyURL') )
  
  ## Showcase demo available live at https://gallery.shinyapps.io/shinyURL
  shinyAppDir( system.file('examples', 'showcase', package='shinyURL') )
  
  ## Interactive R Markdown document which uses a QR code to encode the URL
  if (require("rmarkdown") && require("qrcode"))
    run( system.file('examples', 'qrcode', 'qrcode.Rmd', package='shinyURL') )
    
  ## Use with dynamic user interface created by renderUI()
  shinyAppDir( system.file('examples', 'dynamicUI', package='shinyURL') )
}
}
\author{
Andrzej Oleś <andrzej.oles@embl.de>
}

